stm32: rename mod
