+++ embassy-stm32/src/atomic.rs
