+++ embassy-stm32/src/lib.rs
@@ -16,8 +16,8 @@ mod fmt;
 include!(concat!(env!("OUT_DIR"), "/_macros.rs"));
 
 // Utilities
+mod atomic;
 mod macros;
-mod reg;
 pub mod time;
 mod wait;
 /// Operating modes for peripherals.
